// src/modules/shop/shop.types.ts
export enum ShopActionTypes {
  FETCH_PRODUCTS_BY_TYPE = 'FETCH_PRODUCTS_BY_TYPE',
  FETCH_PRODUCTS_BY_QUERY = 'FETCH_PRODUCTS_BY_QUERY',

  FETCH_PRODUCTS_SUCCESS = 'FETCH_PRODUCTS_SUCCESS',
  FETCH_PRODUCTS_FAILURE = 'FETCH_PRODUCTS_FAILURE',

  FILTER_PRODUCTS = 'FILTER_PRODUCTS',
  INCREASE_LIMIT = 'INCREASE_LIMIT',

  SET_QUERY = 'SET_QUERY',
  CLEAR_QUERY = 'CLEAR_QUERY'
}